blog-backend/requirements.txt
fastapi
uvicorn
pymysql
python-dotenv
pydantic
boto3
python-multipart
